+++ deployment/createGenesis.js
@@ -14,6 +14,8 @@ const {
     MemDB, ZkEVMDB, getPoseidon, smtUtils,
 } = require('@0xpolygonhermez/zkevm-commonjs');
 
+const { deployPolygonZkEVMDeployer, create2Deployment } = require('./helpers/deployment-helpers');
+
 const deployParametersPath = argv.input ? argv.input : './deploy_parameters.json';
 const deployParameters = require(deployParametersPath);
 
@@ -66,13 +68,13 @@ async function main() {
     const proxyAdminFactory = await ethers.getContractFactory('ProxyAdmin', deployer);
     const deployTransactionAdmin = (proxyAdminFactory.getDeployTransaction()).data;
     const dataCallAdmin = proxyAdminFactory.interface.encodeFunctionData('transferOwnership', [deployer.address]);
-    const proxyAdminAddress = await create2Deployment(zkEVMDeployerContract, salt, deployTransactionAdmin, dataCallAdmin);
+    const proxyAdminAddress = await create2Deployment(zkEVMDeployerContract, salt, deployTransactionAdmin, dataCallAdmin, deployer);
 
     // Deploy implementation PolygonZkEVMBridg
     const polygonZkEVMBridgeFactory = await ethers.getContractFactory('PolygonZkEVMBridge', deployer);
     const deployTransactionBridge = (polygonZkEVMBridgeFactory.getDeployTransaction()).data;
     // Mandatory to override the gasLimit since the estimation with create are mess up D:
-    const overrideGasLimit = ethers.BigNumber.from(6000000); // ; // Should be more than enough with 5M
+    const overrideGasLimit = ethers.BigNumber.from(5500000);
     const bridgeImplementationAddress = await create2Deployment(
         zkEVMDeployerContract,
         salt,
@@ -102,7 +104,7 @@ async function main() {
             zkevmAddressL2,
         ],
     );
-    const proxyBridgeAddress = await create2Deployment(zkEVMDeployerContract, salt, deployTransactionProxy, dataCallProxy);
+    const proxyBridgeAddress = await create2Deployment(zkEVMDeployerContract, salt, deployTransactionProxy, dataCallProxy, deployer);
 
     // Import OZ manifest the deployed contracts, its enough to import just the proyx, the rest are imported automatically ( admin/impl)
     await upgrades.forceImport(proxyBridgeAddress, polygonZkEVMBridgeFactory, 'transparent');
@@ -318,95 +320,6 @@ main().catch((e) => {
     console.error(e);
     process.exit(1);
 });
-
-async function deployPolygonZkEVMDeployer(deployer) {
-    const PolgonZKEVMDeployerFactory = await ethers.getContractFactory('PolygonZkEVMDeployer', deployer);
-
-    const deployTxZKEVMDeployer = (PolgonZKEVMDeployerFactory.getDeployTransaction(
-        deployer.address,
-    )).data;
-
-    const gasLimit = ethers.BigNumber.from(1000000); // Put 1 Million, aprox 650k are necessary
-    const gasPrice = ethers.BigNumber.from(ethers.utils.parseUnits('100', 'gwei')); // just in case , seems pretty standard
-    const to = '0x'; // bc deployment transaction, "to" is "0x"
-    const tx = {
-        to,
-        nonce: 0,
-        value: 0,
-        gasLimit: gasLimit.toHexString(),
-        gasPrice: gasPrice.toHexString(),
-        data: deployTxZKEVMDeployer,
-    };
-
-    const signature = {
-        v: 27,
-        r: '0x247000', // Equals 0x0000000000000000000000000000000000000000000000000000000000247000 TODO
-        s: '0x2470', // Equals 0x0000000000000000000000000000000000000000000000000000000000002470 TODO
-    };
-    const serializedTransaction = ethers.utils.serializeTransaction(tx, signature);
-    const resultTransaction = ethers.utils.parseTransaction(serializedTransaction);
-    const totalEther = gasLimit.mul(gasPrice); // 0.1 ether
-
-    // Check if it's already deployed
-    const zkEVMDeployerAddress = ethers.utils.getContractAddress(resultTransaction);
-    if (await ethers.provider.getCode(zkEVMDeployerAddress) !== '0x') {
-        const zkEVMDeployerContract = PolgonZKEVMDeployerFactory.attach(zkEVMDeployerAddress);
-        expect(await zkEVMDeployerContract.owner()).to.be.equal(deployer.address);
-        return [zkEVMDeployerContract, resultTransaction.from];
-    }
-    // Fund keyless deployment
-    const params = {
-        to: resultTransaction.from,
-        value: totalEther.toHexString(),
-    };
-    await (await deployer.sendTransaction(params)).wait();
-
-    // Deploy zkEVMDeployer
-    await (await ethers.provider.sendTransaction(serializedTransaction)).wait();
-
-    const zkEVMDeployerContract = await PolgonZKEVMDeployerFactory.attach(zkEVMDeployerAddress);
-    expect(await zkEVMDeployerContract.owner()).to.be.equal(deployer.address);
-    return [zkEVMDeployerContract, resultTransaction.from];
-}
-
-async function create2Deployment(polgonZKEVMDeployerContract, salt, deployTransaction, dataCall, deployer, hardcodedGasLimit) {
-    // Encode deploy transaction
-    const hashInitCode = ethers.utils.solidityKeccak256(['bytes'], [deployTransaction]);
-
-    // Precalculate create2 address
-    const precalculatedAddressDeployed = ethers.utils.getCreate2Address(polgonZKEVMDeployerContract.address, salt, hashInitCode);
-    const amount = 0;
-
-    if (dataCall) {
-        // Deploy using create2 and call
-        if (hardcodedGasLimit) {
-            const populatedTransaction = await polgonZKEVMDeployerContract.populateTransaction.deployDeterministicAndCall(
-                amount,
-                salt,
-                deployTransaction,
-                dataCall,
-            );
-            populatedTransaction.gasLimit = ethers.BigNumber.from(hardcodedGasLimit);
-            await (await deployer.sendTransaction(populatedTransaction)).wait();
-        } else {
-            await (await polgonZKEVMDeployerContract.deployDeterministicAndCall(amount, salt, deployTransaction, dataCall)).wait();
-        }
-    } else {
-        // Deploy using create2
-        if (hardcodedGasLimit) {
-            const populatedTransaction = await polgonZKEVMDeployerContract.populateTransaction.deployDeterministic(
-                amount,
-                salt,
-                deployTransaction,
-            );
-            populatedTransaction.gasLimit = ethers.BigNumber.from(hardcodedGasLimit);
-            await (await deployer.sendTransaction(populatedTransaction)).wait();
-        } else {
-            await (await polgonZKEVMDeployerContract.deployDeterministic(amount, salt, deployTransaction)).wait();
-        }
-    }
-    return precalculatedAddressDeployed;
-}
 
 async function getAddressInfo(address) {
     const nonce = await ethers.provider.getTransactionCount(address);
