+++ deployment/deployContracts.js
@@ -6,6 +6,8 @@ const path = require('path');
 const fs = require('fs');
 require('dotenv').config({ path: path.resolve(__dirname, '../.env') });
 
+const { deployPolygonZkEVMDeployer, create2Deployment } = require('./helpers/deployment-helpers');
+
 const pathOutputJson = path.join(__dirname, './deploy_output.json');
 const deployParameters = require('./deploy_parameters.json');
 const genesis = require('./genesis.json');
@@ -96,7 +98,7 @@ async function main() {
     console.log('Verifier deployed to:', verifierContract.address);
 
     // Deploy PolygonZkEVMDeployer if is not deployed already using keyless deployment
-    const zkEVMDeployerContract = await deployPolygonZkEVMDeployer(deployer);
+    const [zkEVMDeployerContract, keylessDeployer]  = await deployPolygonZkEVMDeployer(deployer);
 
     /*
      * Deploy Bridge
@@ -107,7 +109,7 @@ async function main() {
     const proxyAdminFactory = await ethers.getContractFactory('ProxyAdmin', deployer);
     const deployTransactionAdmin = (proxyAdminFactory.getDeployTransaction()).data;
     const dataCallAdmin = proxyAdminFactory.interface.encodeFunctionData('transferOwnership', [deployer.address]);
-    const proxyAdminAddress = await create2Deployment(zkEVMDeployerContract, salt, deployTransactionAdmin, dataCallAdmin);
+    const proxyAdminAddress = await create2Deployment(zkEVMDeployerContract, salt, deployTransactionAdmin, dataCallAdmin, deployer);
 
     console.log('#######################\n');
     console.log('Proxy admin deployed to:', proxyAdminAddress);
@@ -116,7 +118,7 @@ async function main() {
     const polygonZkEVMBridgeFactory = await ethers.getContractFactory('PolygonZkEVMBridge', deployer);
     const deployTransactionBridge = (polygonZkEVMBridgeFactory.getDeployTransaction()).data;
     // Mandatory to override the gasLimit since the estimation with create are mess up D:
-    const overrideGasLimit = ethers.BigNumber.from(5500000); // Should be more than enough with 5M
+    const overrideGasLimit = ethers.BigNumber.from(5500000);
     const bridgeImplementationAddress = await create2Deployment(
         zkEVMDeployerContract,
         salt,
@@ -159,7 +161,7 @@ async function main() {
             precalculateZkevmAddress,
         ],
     );
-    const proxyBridgeAddress = await create2Deployment(zkEVMDeployerContract, salt, deployTransactionProxy, dataCallProxy);
+    const proxyBridgeAddress = await create2Deployment(zkEVMDeployerContract, salt, deployTransactionProxy, dataCallProxy, deployer);
     const polygonZkEVMBridgeContract = polygonZkEVMBridgeFactory.attach(proxyBridgeAddress);
 
     console.log('#######################\n');
@@ -379,97 +381,3 @@ main().catch((e) => {
     console.error(e);
     process.exit(1);
 });
-
-async function deployPolygonZkEVMDeployer(deployer) {
-    const PolgonZKEVMDeployerFactory = await ethers.getContractFactory('PolygonZkEVMDeployer', deployer);
-
-    const deployTxZKEVMDeployer = (PolgonZKEVMDeployerFactory.getDeployTransaction(
-        deployer.address,
-    )).data;
-
-    const gasLimit = ethers.BigNumber.from(1000000); // Put 1 Million, aprox 650k are necessary
-    const gasPrice = ethers.BigNumber.from(ethers.utils.parseUnits('100', 'gwei')); // just in case , seems pretty standard
-    const to = '0x'; // bc deployment transaction, "to" is "0x"
-    const tx = {
-        to,
-        nonce: 0,
-        value: 0,
-        gasLimit: gasLimit.toHexString(),
-        gasPrice: gasPrice.toHexString(),
-        data: deployTxZKEVMDeployer,
-    };
-
-    const signature = {
-        v: 27,
-        r: '0x247000', // Equals 0x0000000000000000000000000000000000000000000000000000000000247000 TODO
-        s: '0x2470', // Equals 0x0000000000000000000000000000000000000000000000000000000000002470 TODO
-    };
-    const serializedTransaction = ethers.utils.serializeTransaction(tx, signature);
-    const resultTransaction = ethers.utils.parseTransaction(serializedTransaction);
-    const totalEther = gasLimit.mul(gasPrice); // 0.1 ether
-
-    // Check if it's already deployed
-    const zkEVMDeployerAddress = ethers.utils.getContractAddress(resultTransaction);
-    if (await ethers.provider.getCode(zkEVMDeployerAddress) !== '0x') {
-        const zkEVMDeployerContract = PolgonZKEVMDeployerFactory.attach(zkEVMDeployerAddress);
-        expect(await zkEVMDeployerContract.owner()).to.be.equal(deployer.address);
-        return zkEVMDeployerContract;
-    }
-
-    // Fund keyless deployment
-    const params = {
-        to: resultTransaction.from,
-        value: totalEther.toHexString(),
-    };
-    await (await deployer.sendTransaction(params)).wait();
-
-    // Deploy zkEVMDeployer
-    await (await ethers.provider.sendTransaction(serializedTransaction)).wait();
-
-    const zkEVMDeployerContract = await PolgonZKEVMDeployerFactory.attach(zkEVMDeployerAddress);
-    expect(await zkEVMDeployerContract.owner()).to.be.equal(deployer.address);
-    return zkEVMDeployerContract;
-}
-
-async function create2Deployment(polgonZKEVMDeployerContract, salt, deployTransaction, dataCall, deployer, hardcodedGasLimit) {
-    // Encode deploy transaction
-    const hashInitCode = ethers.utils.solidityKeccak256(['bytes'], [deployTransaction]);
-
-    // Precalculate create2 address
-    const precalculatedAddressDeployed = ethers.utils.getCreate2Address(polgonZKEVMDeployerContract.address, salt, hashInitCode);
-    const amount = 0;
-
-    if (await ethers.provider.getCode(precalculatedAddressDeployed) !== '0x') {
-        return precalculatedAddressDeployed;
-    }
-
-    if (dataCall) {
-        // Deploy using create2 and call
-        if (hardcodedGasLimit) {
-            const populatedTransaction = await polgonZKEVMDeployerContract.populateTransaction.deployDeterministicAndCall(
-                amount,
-                salt,
-                deployTransaction,
-                dataCall,
-            );
-            populatedTransaction.gasLimit = ethers.BigNumber.from(hardcodedGasLimit);
-            await (await deployer.sendTransaction(populatedTransaction)).wait();
-        } else {
-            await (await polgonZKEVMDeployerContract.deployDeterministicAndCall(amount, salt, deployTransaction, dataCall)).wait();
-        }
-    } else {
-        // Deploy using create2
-        if (hardcodedGasLimit) {
-            const populatedTransaction = await polgonZKEVMDeployerContract.populateTransaction.deployDeterministic(
-                amount,
-                salt,
-                deployTransaction,
-            );
-            populatedTransaction.gasLimit = ethers.BigNumber.from(hardcodedGasLimit);
-            await (await deployer.sendTransaction(populatedTransaction)).wait();
-        } else {
-            await (await polgonZKEVMDeployerContract.deployDeterministic(amount, salt, deployTransaction)).wait();
-        }
-    }
-    return precalculatedAddressDeployed;
-}
