+++ index.js
@@ -3,7 +3,7 @@ module.exports.PolygonZkEVMGlobalExitRoot = require('./compiled-contracts/Polygo
 module.exports.PolygonZkEVMGlobalExitRootL2 = require('./compiled-contracts/PolygonZkEVMGlobalExitRootL2.json');
 module.exports.PolygonZkEVM = require('./compiled-contracts/PolygonZkEVM.json');
 module.exports.TokenWrapped = require('./compiled-contracts/TokenWrapped.json');
-module.exports.Verifier = require('./compiled-contracts/Verifier.json');
+module.exports.FflonkVerifier = require('./compiled-contracts/FflonkVerifier.json');
 module.exports.PolygonZkEVMBridgeMock = require('./compiled-contracts/PolygonZkEVMBridgeMock.json');
 module.exports.ERC20PermitMock = require('./compiled-contracts/ERC20PermitMock.json');
 module.exports.PolygonZkEVMGlobalExitRootL2Mock = require('./compiled-contracts/PolygonZkEVMGlobalExitRootL2Mock.json');
